+++ apps/omg_child_chain/priv/test_fees_file--576460741410378700
@@ -1,0 +1,1 @@
+{"1":{"0x0000000000000000000000000000000000000000":{"amount":1,"pegged_amount":1,"pegged_currency":"USD","pegged_subunit_to_unit":100,"subunit_to_unit":1000000000000000000,"symbol":"token","type":"fixed","updated_at":"2020-05-26T14:50:22.090452Z"},"0xdd2860dd8f182f90870383a98ddaf63fdb00573e":{"amount":2,"pegged_amount":1,"pegged_currency":"USD","pegged_subunit_to_unit":100,"subunit_to_unit":1000000000000000000,"symbol":"token","type":"fixed","updated_at":"2020-05-26T14:50:22.090465Z"}}}
