+++ apps/omg_watcher/test/omg_watcher/integration/in_flight_exit_test.exs
@@ -218,9 +218,9 @@ defmodule OMG.Watcher.Integration.InFlightExitTest do
     IntegrationTest.wait_for_block_fetch(blknum, @timeout)
 
     _ = exit_in_flight_and_wait_for_ife(ife1, alice)
-
+   
     assert %{"in_flight_exits" => [%{}]} = WatcherHelper.success?("/status.get")
-
+    
     _ = piggyback_and_process_exits(tx, 1, :output, bob)
 
     assert %{"in_flight_exits" => [], "byzantine_events" => []} = WatcherHelper.success?("/status.get")
